% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCoastline.R
\name{addCoastline}
\alias{addCoastline}
\title{Add Coastline to basemap}
\usage{
addCoastline(
  basemap,
  coastline = NULL,
  land.col = "lightgrey",
  lons = c(-180, 180),
  lats = c(-90, 90)
)
}
\description{
Adds a coastline list to a map
}
\author{
Thomas Bryce Kelly
}
