% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEEZ.R
\name{addEEZ}
\alias{addEEZ}
\title{Add EEZ to map}
\usage{
addEEZ(basemap, eez = NULL, pattern = "*", ..., verbose = T)
}
\arguments{
\item{basemap}{a list object such as from plotBasemap()}

\item{eez}{a list of eez coordinates such as that provided herein as eez5.}

\item{verbose}{a boolean flag to turn on/off displayed messages}
}
\description{
Add the EEZ boundary lines to a map
}
\author{
Thomas Bryce Kelly
}
