% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLatitude.R
\name{addLatitude}
\alias{addLatitude}
\title{Add Latitudinal Graticules to Map}
\usage{
addLatitude(basemap, lats = NULL, sides = 2, col = "#00000030", ...)
}
\arguments{
\item{basemap}{basemap such as that from plotBasemap()}

\item{lats}{vector of latitude values (in degree northing) to overlay}

\item{sides}{integer or vector of integers for which sides to add axes (if any).}

\item{col}{color value for the graticule lines}

\item{...}{optional arguments passed on to lines()}
}
\description{
Add Latitudinal Graticules to Map
}
\author{
Thomas Bryce Kelly
}
