% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLayer.R
\name{addLayer}
\alias{addLayer}
\title{Add Map Layer}
\usage{
addLayer(basemap, lon, lat, z, zlim = NULL, pal = greyscale(255), verbose = T)
}
\description{
Add a image layer to the map!
}
\author{
Thomas Bryce Kelly
}
