% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLongitude.R
\name{addLongitude}
\alias{addLongitude}
\title{Add Longitudinal Graticules to Map}
\usage{
addLongitude(basemap, lons = NULL, sides = 1, col = "#00000030", ...)
}
\arguments{
\item{basemap}{basemap such as that from plotBasemap()}

\item{lons}{vector of longitude values (in degree easting) to overlayt}

\item{sides}{integer or vector of integers for which sides to add axes (if any).}

\item{col}{color value for the graticule lines}

\item{...}{optional arguments passed on to lines()}
}
\description{
Add Longitudinal Graticules to Map
}
\author{
Thomas Bryce Kelly
}
