% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPoints.R
\name{addPoints}
\alias{addPoints}
\title{Add Map Points}
\usage{
addPoints(basemap, lon, lat, col = "black", pch = 1, cex = 1, ...)
}
\arguments{
\item{basemap}{a basemap list object}

\item{lon}{a set of longitudes to draw a line between}

\item{lat}{a set of latitudes to draw a lien between}

\item{col}{the color of the line to be drawn}

\item{pch}{pch}

\item{cex}{size}

\item{...}{passed onto points()}
}
\description{
Add Map Points
}
\author{
Laura Whitmore
}
