% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.vertex.R
\name{calc.vertex}
\alias{calc.vertex}
\title{Calculate extended grid}
\usage{
calc.vertex(x, y)
}
\arguments{
\item{x}{x positiosn of the grid}

\item{y}{y positions of the grid}
}
\description{
Calculate extended grid
}
\author{
Thomas Bryce Kelly
}
