% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceKM.R
\name{distanceKM}
\alias{distanceKM}
\title{Calculate Distance between Points (kilometers)}
\usage{
distanceKM(lon1, lat1, lon2, lat2)
}
\arguments{
\item{lon1}{longitude of first point}

\item{lat1}{latitude of first point}

\item{lon2}{longitude of second point}

\item{lat2}{latitude of second point}
}
\description{
Calculate Distance between Points (kilometers)
}
