% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIntersection.R
\name{getIntersection}
\alias{getIntersection}
\title{Utility Func}
\usage{
getIntersection(x, y)
}
\arguments{
\item{x}{x coordinates of a line}

\item{y}{y coordinates of a line}
}
\description{
Utility Func
}
\author{
Thomas Bryce Kelly
}
