% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSubsample.R
\name{gridSubsample}
\alias{gridSubsample}
\title{Calculate subsampled grid}
\usage{
gridSubsample(x = NULL, y = NULL, z, approx = F)
}
\arguments{
\item{x}{the x positions in the grid}

\item{y}{the y positions in the grid}

\item{z}{the grid of values to be refined}

\item{approx}{boolean flag to turn on fast-approximation method}
}
\description{
Calculate subsampled grid
}
\author{
Thomas Bryce Kelly
}
