% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBasemap.R
\name{plotBasemap}
\alias{plotBasemap}
\title{Make Map}
\usage{
plotBasemap(
  coastline = NULL,
  lon = 0,
  lat = 0,
  scale = 1000,
  projection = NULL,
  land.col = "lightgray",
  verbose = T
)
}
\arguments{
\item{land.col}{A color string or value for the land polygon}

\item{coast}{Should be the name of a coastline data object. A value of NULL sets the default cosatline to 'coastline2'.}
}
\description{
Make Map
}
\author{
Laura Whitmore

Thomas Bryce Kelly
}
