% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectEqualarea.R
\name{projectionEqualarea}
\alias{projectionEqualarea}
\title{Equal Area Projection}
\usage{
projectionEqualarea(lon, lat, lon0 = 0, lat0 = 0, inv = F)
}
\arguments{
\item{lon}{longitude values to project (or x values when inverting)}

\item{lat}{latitude values to project (or y values when inverting)}

\item{lon0}{center longitude value of the projection}

\item{lat0}{center latitude value of the projection}

\item{inv}{a boolean flag to turn on/off projection inversion}
}
\description{
Equal Area Projection
}
