% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectOrthographic.R
\name{projectionOrthographic}
\alias{projectionOrthographic}
\title{Orthographic Projection}
\usage{
projectionOrthographic(lon, lat, lon0 = 0, lat0 = 0, inv = F)
}
\description{
Orthographic Projection
}
