% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionStereographic.R
\name{projectionStereographic}
\alias{projectionStereographic}
\title{Stereographic Projection}
\usage{
projectionStereographic(lon, lat, lon0 = 0, lat0 = 0, inv = F)
}
\description{
Stereographic Projection
}
