% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replot.R
\name{replot}
\alias{replot}
\title{Replot a previous map}
\usage{
replot(basemap)
}
\arguments{
\item{basemap}{a list structure generated via plotBasemap()}
}
\description{
A simple function that allows multiple plotting commands to be replayed from a previously rendered map.
}
