% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitBool.R
\name{splitBool}
\alias{splitBool}
\title{Boolean Vector Split
A helper function that takes a boolean vector and returns a list of indicies for each sequence of consecutive T/F values.}
\usage{
splitBool(vec)
}
\arguments{
\item{vec}{a boolean vector to be split}
}
\description{
Boolean Vector Split
A helper function that takes a boolean vector and returns a list of indicies for each sequence of consecutive T/F values.
}
\author{
Thomas Bryce Kelly
}
