% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.longitude.R
\name{standardize.longitude}
\alias{standardize.longitude}
\title{Standardize longitude -180 - 180}
\usage{
standardize.longitude(lon)
}
\arguments{
\item{lon}{longitude values to be standardize into [-180, +180]}
}
\description{
Standardize longitude -180 - 180
}
