% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCoastline.R
\name{addCoastline}
\alias{addCoastline}
\title{Add Coastline to basemap}
\usage{
addCoastline(basemap, coastline, land.col)
}
\description{
Adds a coastline list to a map given a projection and land color.
}
\author{
Thomas Bryce Kelly
}
