% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRadius.R
\name{addRadius}
\alias{addRadius}
\title{Add reference radius to a map}
\usage{
addRadius(basemap, lon, lat, radius, ...)
}
\arguments{
\item{basemap}{a list object generated from plotBasemap}

\item{lon}{longitude}

\item{lat}{latitude}

\item{radius}{radius of circle in km}

\item{...}{additional arguments to be passed into addLine}
}
\description{
Add reference radius to a map
}
