% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbar.R
\name{colorbar}
\alias{colorbar}
\title{All Colorbar}
\usage{
colorbar(
  pal,
  zlim,
  num_labels = NULL,
  x = NULL,
  y = NULL,
  width = 1,
  height = 1
)
}
\description{
All Colorbar
}
\author{
Thomas Bryce Kelly
}
\keyword{Gridding}
