% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldOfView.R
\name{fieldOfView}
\alias{fieldOfView}
\title{Add Coastline to basemap}
\usage{
fieldOfView(basemap, n = 10)
}
\arguments{
\item{basemap}{a list generated from plotBasemap()}

\item{n}{an integer used to set the resolution of the FOV grid.}
}
\description{
Adds a coastline list to a map
}
\author{
Thomas Bryce Kelly
}
