% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSubsample.R
\name{gridSubsample}
\alias{gridSubsample}
\title{Calculate subsampled grid}
\usage{
gridSubsample(x = NULL, y = NULL, z, approx = F)
}
\description{
Calculate subsampled grid
}
\author{
Thomas Bryce Kelly
}
