% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBasemap.R
\name{plotBasemap}
\alias{plotBasemap}
\title{Make Map}
\usage{
plotBasemap(
  lon = 0,
  lat = 0,
  scale = 1000,
  coastline = NULL,
  projection = projectionEqualarea,
  land.col = "lightgray",
  frame = T,
  verbose = T
)
}
\arguments{
\item{lon}{center longitude of map}

\item{lat}{center latitude of map}

\item{scale}{the visual scale of the map (in km)}

\item{coastline}{a coastline data list object (optional)}

\item{projection}{a function used to project (and invert the projection)}

\item{land.col}{color used for the fill of each land polygon}

\item{frame}{a boolean flag to redraw the graphic canvas' frame}

\item{verbose}{a flag to turn on/off pritning to screen}
}
\description{
Make Map
}
\author{
Laura Whitmore

Thomas Bryce Kelly
}
